+++ addons/fr/choruspro/bill.go
@@ -1,6 +1,8 @@
 package choruspro
 
 import (
+	"fmt"
+
 	"github.com/invopop/gobl/bill"
 	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/tax"
@@ -11,22 +13,22 @@ import (
 func validateInvoice(inv *bill.Invoice) error {
 	return validation.ValidateStruct(inv,
 		validation.Field(&inv.Customer,
-			validation.By(validateCustomer),
+			validation.By(validateInvoiceCustomer),
 		),
 		validation.Field(&inv.Tax,
 			validation.Required,
-			validation.By(validateTax),
+			validation.By(validateInvoiceTax),
 		),
 		validation.Field(&inv.Totals,
 			validation.When(
 				// A2 can only exist if invoice has been paid
-				inv.Tax != nil && inv.Tax.Ext != nil && inv.Tax.Ext.Get(ExtKeyFramework) == "A2",
-				validation.By(validatePaid),
+				inv.Tax != nil && inv.Tax.Ext.Get(ExtKeyFramework) == ExtFrameworkCodePaid,
+				validation.By(validateInvoicePaid),
 			)),
 	)
 }
 
-func validateCustomer(value interface{}) error {
+func validateInvoiceCustomer(value interface{}) error {
 	customer, ok := value.(*org.Party)
 	if !ok || customer == nil {
 		return nil
@@ -44,7 +46,7 @@ func validateCustomer(value interface{}) error {
 	)
 }
 
-func validateTax(value interface{}) error {
+func validateInvoiceTax(value interface{}) error {
 	t, ok := value.(*bill.Tax)
 	if !ok || t == nil {
 		return nil
@@ -72,24 +74,26 @@ func normalizeInvoice(inv *bill.Invoice) {
 		inv.Tax.Ext = make(tax.Extensions)
 	}
 
-	// Set default framework type if not specified
+	// Set default framework type if not specified. This breaks away from the
+	// typical deterministic behavior of assigning extensions in GOBL, due to
+	// complexity of trying to apply scenarios.
 	if !inv.Tax.Ext.Has(ExtKeyFramework) {
 		inv.Tax.Ext = inv.Tax.Ext.Merge(
 			tax.Extensions{
-				ExtKeyFramework: "A1",
+				ExtKeyFramework: ExtFrameworkCodeSupplier,
 			},
 		)
 	}
 
 }
 
-func validatePaid(value interface{}) error {
+func validateInvoicePaid(value interface{}) error {
 	totals, ok := value.(*bill.Totals)
 	if !ok {
 		return nil
 	}
 	if !totals.Paid() {
-		return validation.NewError("totals", "If the invoice has type A2, it must be paid in full")
+		return fmt.Errorf("must be paid in full for framework '%s'", ExtFrameworkCodePaid)
 	}
 	return nil
 }