+++ CHANGELOG.md
@@ -6,11 +6,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Changed
+
+- `es-verifactu-v1`: `[]*org.Note` validation will check the note's length if `key=general`
+
 ### Added
 
 - `eu-en16931-v2017`: Add missing Business Rules with labels at implementation level.
 - `de-xrechnung-v3`: Add missing validation from CIUS.
 - `fr`: Add identity validation and normalization.
+- `it-sdi-v1`: Validate names and addresses contain only latin characters as required by SDI.
+- `it-sdi-v1`: Payment means now support `online` key as a card payment.
+- `org`: `Unit` additional standard units.
 - `es/verifactu`: map identity keys for specific codes.
 - `org`: `Person` new `key`, `addresses`, and `identities` properties.